Add files via upload
In order to run this application you need to install electron.

npm instal electron

This should do the trick. Once you run the application (npm start) the window opens and you can play. Every cellmarking sends a POST message to the "toilet" in JSON format. You find the link to the "toilet" (so you can see what you sent) in the code as a comment.

Babystep after babystep we get closer to finishing this damn project.